+++ src/components/tabs/VCSidebarTabs.tsx
@@ -1,40 +1,46 @@
 import React, { useEffect, useState } from 'react';
 import { TbFileMusic } from 'react-icons/tb';
 
-import KeBob from '@/components/dropdowns/kebob';
+import KeBob from '@/components/dropdowns/KebabMenu';
 import FileUploadBox from '@/components/feature/FileUploadBox';
 import Pagination from '@/components/ui/pagination';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
-interface VCselect {
+
+interface VCSidebarTabs {
   id: string;
   name: string;
   size: number;
   isEditing?: boolean;
 }
-const dummyFiles: VCselect[] = Array.from({ length: 8 }, (_, i) => ({
+
+const dummyFiles: VCSidebarTabs[] = Array.from({ length: 8 }, (_, i) => ({
   id: `file-${i + 1}`,
   name: `샘플 보이스 ${i + 1}.wav`,
   size: 104.2 * 1024 * 1024,
 }));
-const VCselectTabs = () => {
-  const [activeFile, setActiveFile] = useState<VCselect | null>(null);
-  const [files, setFiles] = useState<VCselect[]>(dummyFiles);
+
+const VCSidebarTabs = () => {
+  const [activeFile, setActiveFile] = useState<VCSidebarTabs | null>(null);
+  const [files, setFiles] = useState<VCSidebarTabs[]>(dummyFiles);
   const [currentPage, setCurrentPage] = useState(1);
   const filesPerPage = 4;
   const formatFileSize = (size: number) => {
     const sizeInMB = size / (1024 * 1024);
     return sizeInMB < 0.1 ? '0.1' : sizeInMB.toFixed(1);
   };
-  const handleFileUpload = (newFile: VCselect) => {
+
+  const handleFileUpload = (newFile: VCSidebarTabs) => {
     setFiles((prev) => {
       const isDuplicate = prev.some((file) => file.name === newFile.name);
       if (isDuplicate) return prev;
       return [newFile, ...prev];
     });
   };
-  const handleFileSelect = (file: VCselect) => {
+
+  const handleFileSelect = (file: VCSidebarTabs) => {
     setActiveFile(file);
   };
+
   const handleStartRename = (fileId: string) => {
     setFiles(
       files.map((file) => ({
@@ -43,6 +49,7 @@ const VCselectTabs = () => {
       }))
     );
   };
+
   const handleRename = (fileId: string, event: React.KeyboardEvent<HTMLInputElement>) => {
     if (event.key === 'Enter') {
       const newName = (event.target as HTMLInputElement).value;
@@ -58,6 +65,7 @@ const VCselectTabs = () => {
       }
     }
   };
+
   const handleDelete = (fileId: string) => {
     setFiles(files.filter((file) => file.id !== fileId));
     if (activeFile?.id === fileId) {
@@ -71,7 +79,7 @@ const VCselectTabs = () => {
   }, [files.length]);
   const totalPages = Math.ceil(files.length / filesPerPage);
   const currentFiles = files.slice((currentPage - 1) * filesPerPage, currentPage * filesPerPage);
-  const FileListItem = ({ file }: { file: VCselect }) => (
+  const FileListItem = ({ file }: { file: VCSidebarTabs }) => (
     <div
       className={`flex items-center justify-between h-[61px] px-3
         border-b border-gray-300 bg-background cursor-pointer
@@ -152,4 +160,4 @@ const VCselectTabs = () => {
     </div>
   );
 };
-export default VCselectTabs;
+export default VCSidebarTabs;
